Big update: New ORAM interface with runtime selection Internal revisions to Queue and other components Added batcher sort (no unit tests yet) Started work on new Gale-Shapley implementation
Currently broken:
Circuit ORAM segfaults
Gale Shapley returns incorrect results (and segfaults, but only sometimes)
Building the benchmarks is disabled

+++ src/uint128.c
@@ -38,7 +38,7 @@ uint128_t * uint128_new() {
     return result;
 }
 
-uint128_t * uint128_from(uint64_t input) {
+uint128_t * uint128_from(unsigned __int128 input) {
     uint128_t * result = malloc(sizeof(uint128_t));
     result->val = input;
     return result;
