+++ src/uint128.h
@@ -8,7 +8,7 @@ typedef struct uint128_t uint128_t;
 uint128_t * uint128_current_timestamp();
 
 uint128_t * uint128_new();
-uint128_t * uint128_from(uint64_t);
+uint128_t * uint128_from(unsigned __int128);
 uint128_t * uint128_add(uint128_t *, uint128_t *);
 uint128_t * uint128_subtract(uint128_t *, uint128_t *);
 uint128_t * uint128_multiply(uint128_t *, uint128_t *);
